Added cors, make some components in client, wrote some client services, make routing, models

+++ AgendaCalendar.Application/Reminders/Commands/DeleteReminderCommand.cs
@@ -1,13 +1,17 @@
 ﻿
 namespace AgendaCalendar.Application.Reminders.Commands
 {
-    public sealed record DeleteReminderCommand(int reminderId) : IRequest<Reminder> { }
+    public sealed record DeleteReminderCommand(int reminderId) : IRequest<ErrorOr<Reminder>> { }
 
-    public class DeleteReminderCommandHandler(IUnitOfWork unitOfWork) : IRequestHandler<DeleteReminderCommand, Reminder>
+    public class DeleteReminderCommandHandler(IUnitOfWork unitOfWork) : IRequestHandler<DeleteReminderCommand, ErrorOr<Reminder>>
     {
-        public async Task<Reminder> Handle(DeleteReminderCommand request, CancellationToken cancellationToken)
+        public async Task<ErrorOr<Reminder>> Handle(DeleteReminderCommand request, CancellationToken cancellationToken)
         {
             var reminder = await unitOfWork.ReminderRepository.GetByIdAsync(request.reminderId);
+            if(reminder is null)
+            {
+                return Errors.Reminder.NotFound;
+            }
 
             await unitOfWork.ReminderRepository.DeleteAsync(reminder.Id);
             await unitOfWork.SaveAllAsync();
